Added temperature warning to OSD and started implementing battery warning

+++ Aerovelo/HPV_OSD/OSD_SLIP.cpp
@@ -14,7 +14,7 @@ extern char profileName[16];
 extern uint8_t calibrationState, mode;
 extern int16_t offset;
 extern float batteryLevel;
-extern int8_t temperature;
+extern int16_t temperature;
 extern uint8_t GPSComm, SDComm;
 
 
@@ -102,7 +102,7 @@ void OSD_SlipParse(char *slipBuffer) {
     batteryLevel = *((uint16_t*)((uint8_t*)slipBuffer+1)) * 1.0 / 100;
     break;
   case ID_TEMPERATURE:
-    temperature = *((int8_t*) (slipBuffer+1));
+    temperature = *((int16_t*) (slipBuffer+1));
     break;
   case ID_GPSCOMM:
     GPSComm = *((uint8_t*) (slipBuffer+1));
